Update image exists check to use manifest endpoint
With the new version update of docker registry, pagination was
introduced. This is not supported by the docker-registry-client package
at this point. We now check if the requested tag's manifest can be
requested as a check for its existence. This is a more optimal
solution than listing all tags for a repository.